# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Patient {
  weight_kg: number;
  age_years: number;
  allergies?: string[];
  current_medications?: string[];
  conditions?: string[];
  renal_impairment?: boolean;
}

export interface DrugSummary {
  name: string;
  display_name: string;
  drug_class: string;
  indication: string;
  routes: string[];
  tags: string[];
  /** @nullable */
  controlled_substance?: string | null;
  /** @nullable */
  reversal_agent?: string | null;
}

export interface DrugDetail {
  name: string;
  display_name: string;
  brand_names: string[];
  drug_class: string;
  indication: string;
  mechanism: string;
  routes: string[];
  contraindications: string[];
  pregnancy_category: string;
  /** @nullable */
  controlled_substance?: string | null;
  renal_caution: boolean;
  hepatic_caution: boolean;
  /** @nullable */
  reversal_agent?: string | null;
  tags: string[];
}

export interface DrugStats {
  total_drugs: number;
  drug_classes: string[];
  top_tags: string[];
  controlled_count: number;
  reversal_agent_count: number;
}

export interface DoseRequest {
  drug_name: string;
  patient: Patient;
  route?: string;
  dose_fraction?: number;
}

export interface ContraindicationFlag {
  absolute: boolean;
  detail: string;
}

export interface InteractionFlag {
  drug_a: string;
  drug_b: string;
  severity: string;
  description: string;
  management: string;
}

export interface DoseResult {
  summary: string;
  drug_name: string;
  display_name: string;
  route: string;
  dose_display: number;
  dose_unit: string;
  volume_ml: number;
  frequency: string;
  contraindications: ContraindicationFlag[];
  interactions: InteractionFlag[];
}

export interface DripRequest {
  drug_name: string;
  patient: Patient;
  ordered_dose: number;
  dose_unit: string;
  /** @nullable */
  concentration_mg_ml?: number | null;
  bag_volume_ml?: number;
}

export interface DripResult {
  summary: string;
  drug_name: string;
  display_name: string;
  rate_ml_per_hr: number;
  rate_ml_per_min: number;
  /** @nullable */
  duration_hr?: number | null;
  bag_volume_ml: number;
  concentration_mg_ml: number;
  warnings: string[];
  rate_exceeded: boolean;
}

export interface VitalsRequest {
  respiratory_rate: number;
  spo2: number;
  on_oxygen: boolean;
  systolic_bp: number;
  heart_rate: number;
  /** ACVPU level: A, C, V, P, or U */
  consciousness: string;
  temperature: number;
  /** @nullable */
  gcs_eye?: number | null;
  /** @nullable */
  gcs_verbal?: number | null;
  /** @nullable */
  gcs_motor?: number | null;
}

export interface News2Score {
  total: number;
  risk: string;
  interpretation: string;
}

export interface GcsScore {
  total: number;
  severity: string;
  interpretation: string;
  eye_descriptor: string;
  verbal_descriptor: string;
  motor_descriptor: string;
}

export interface QsofaScore {
  total: number;
  sepsis_alert: boolean;
  interpretation: string;
}

export interface VitalsAssessment {
  news2: News2Score;
  gcs?: GcsScore;
  qsofa: QsofaScore;
}

export interface SafetyRequest {
  drug_name: string;
  patient: Patient;
  /** @nullable */
  route?: string | null;
}

export interface SafetyResult {
  drug_name: string;
  display_name: string;
  contraindications: ContraindicationFlag[];
  interactions: InteractionFlag[];
  has_absolute_contraindication: boolean;
  has_major_interaction: boolean;
}

export type SearchDrugsParams = {
  q?: string;
  tag?: string;
};