# lib/api-zod/src/generated/types/contraindicationFlag.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */

export interface ContraindicationFlag {
  absolute: boolean;
  detail: string;
}