# lib/api-zod/src/generated/types/doseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */
import type { Patient } from "./patient";

export interface DoseRequest {
  drug_name: string;
  patient: Patient;
  route?: string;
  dose_fraction?: number;
}