# lib/api-zod/src/generated/types/doseResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */
import type { ContraindicationFlag } from "./contraindicationFlag";
import type { InteractionFlag } from "./interactionFlag";

export interface DoseResult {
  summary: string;
  drug_name: string;
  display_name: string;
  route: string;
  dose_display: number;
  dose_unit: string;
  volume_ml: number;
  frequency: string;
  contraindications: ContraindicationFlag[];
  interactions: InteractionFlag[];
}