# lib/api-zod/src/generated/types/dripResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */

export interface DripResult {
  summary: string;
  drug_name: string;
  display_name: string;
  rate_ml_per_hr: number;
  rate_ml_per_min: number;
  /** @nullable */
  duration_hr?: number | null;
  bag_volume_ml: number;
  concentration_mg_ml: number;
  warnings: string[];
  rate_exceeded: boolean;
}