# lib/api-zod/src/generated/types/drugDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */

export interface DrugDetail {
  name: string;
  display_name: string;
  brand_names: string[];
  drug_class: string;
  indication: string;
  mechanism: string;
  routes: string[];
  contraindications: string[];
  pregnancy_category: string;
  /** @nullable */
  controlled_substance?: string | null;
  renal_caution: boolean;
  hepatic_caution: boolean;
  /** @nullable */
  reversal_agent?: string | null;
  tags: string[];
}