# lib/api-zod/src/generated/types/drugStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */

export interface DrugStats {
  total_drugs: number;
  drug_classes: string[];
  top_tags: string[];
  controlled_count: number;
  reversal_agent_count: number;
}