# lib/api-zod/src/generated/types/drugSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */

export interface DrugSummary {
  name: string;
  display_name: string;
  drug_class: string;
  indication: string;
  routes: string[];
  tags: string[];
  /** @nullable */
  controlled_substance?: string | null;
  /** @nullable */
  reversal_agent?: string | null;
}