# lib/api-zod/src/generated/types/gcsScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */

export interface GcsScore {
  total: number;
  severity: string;
  interpretation: string;
  eye_descriptor: string;
  verbal_descriptor: string;
  motor_descriptor: string;
}