# lib/api-zod/src/generated/types/interactionFlag.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */

export interface InteractionFlag {
  drug_a: string;
  drug_b: string;
  severity: string;
  description: string;
  management: string;
}