# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  weight_kg: number;
  age_years: number;
  allergies?: string[];
  current_medications?: string[];
  conditions?: string[];
  renal_impairment?: boolean;
}