# lib/api-zod/src/generated/types/qsofaScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */

export interface QsofaScore {
  total: number;
  sepsis_alert: boolean;
  interpretation: string;
}