# lib/api-zod/src/generated/types/safetyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */
import type { Patient } from "./patient";

export interface SafetyRequest {
  drug_name: string;
  patient: Patient;
  /** @nullable */
  route?: string | null;
}