# lib/api-zod/src/generated/types/safetyResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinical Decision Support API
 * OpenAPI spec version: 0.1.0
 */
import type { ContraindicationFlag } from "./contraindicationFlag";
import type { InteractionFlag } from "./interactionFlag";

export interface SafetyResult {
  drug_name: string;
  display_name: string;
  contraindications: ContraindicationFlag[];
  interactions: InteractionFlag[];
  has_absolute_contraindication: boolean;
  has_major_interaction: boolean;
}